refactor(core): inline single integration protocol (R16)
